+++ src/gui/widget/toolbar.cpp
@@ -58,7 +58,7 @@ namespace neogfx
 
     uint32_t toolbar::button_count() const
     {
-        return iButtons.size();
+        return static_cast<uint32_t>(iButtons.size());
     }
 
     const toolbar_button& toolbar::button(button_index aButtonIndex) const
